requirements.txt
google-api-python-client>=2.100.0
requests>=2.32.4
aiohttp>=3.8.0
psycopg2-binary>=2.9.0